Flask_Api/requirements.txt
Flask==3.1.0
numpy==1.23.5
scikit-learn==1.2.2
pandas==2.2.3
tensorflow==2.17.1
pillow==11.0.0
flask_cors

